Incorporate a minimalistic lighting model to the scene
This allows me to better look at normals and if they are being properly calculated. I have to make this model better though, right now is basically directional lighting (simulating a very distant light source, like the sun, where distance doesn't factor).

There is a lot of renaming and cleanup on this commit as well.